app/backend/requirements.txt
fastapi~=0.115.11
uvicorn
openai~=1.68.2
supabase~=2.14.0
python-multipart
pypdf2~=3.0.1
python-docx~=1.1.2
pydantic~=2.10.6